requirements.txt
streamlit>=1.35
groq>=0.9
faiss-cpu>=1.8
sentence-transformers>=3.0
numpy>=1.26
pandas>=2.0